Moved SessionTimeView to common component.
